Move tour and demo to the tour app
