+++ pontoon/tour/static/css/tour.css
