+++ pontoon/tour/static/js/lib/sideshow/fonts/sideshow-icons/sideshow-icons.eot
