+++ pontoon/tour/static/js/tour.config.js
