+++ pontoon/tour/static/js/tour.js
