+++ pontoon/urls.py
@@ -102,8 +102,9 @@ urlpatterns = [
     url(r'^static/js/pontoon\.js$', pontoon_js_view),
 
     # Include URL configurations from installed apps
-    url(r'', include('pontoon.tags.urls')),
     url(r'', include('pontoon.teams.urls')),
+    url(r'', include('pontoon.tour.urls')),
+    url(r'', include('pontoon.tags.urls')),
     url(r'', include('pontoon.sync.urls')),
     url(r'', include('pontoon.projects.urls')),
     url(r'', include('pontoon.machinery.urls')),
